ENH: Delay setup until after scheduler setup in run_espei.py
When using MPI, we start a version of ESPEI for each MPI task, which gets
expensive for startup time. Moving anything that can come after making
all of the non-master MPI tasks wait should improve the speed.
Specifically, we moved the tracefile, probfile, and datasets loading
and checking after waiting on all on the slave MPI tasks.